requirements.txt
Flask
Flask-Script
Flask-SQLAlchemy
Flask-Migrate
Flask-Classy
Flask_Redis
mockredispy
gunicorn
pymysql
